Move AudioIOExt to its own files
