+++ src/MIDIPlay.cpp
@@ -356,6 +356,7 @@ Time (in seconds, = total_sample_count / sample_rate)
 
 #include "MIDIPlay.h"
 #include "AudioIO.h"
+
 #include "Prefs.h"
 #include "portaudio.h"
 #include <portmidi.h>
